:recycle: Remove blockchain prefix
