+++ packages/lisk-p2p/src/index.ts
@@ -12,4 +12,4 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-export * from './blockchain';
+export * from './blockchain-p2p';
